Tidy imports - icons and contextMenu moved to subfolders
