+++ components/InputContainer.tsx
@@ -1,6 +1,19 @@
 import styled, {css} from 'styled-components';
 import { PALETTE, LAYOUT, SHADOW, TYPOGRAPHY } from '../utils/Theme';
 
+// export const StyledLabel = styled.label<{disabled : boolean}>`
+//     ${TYPOGRAPHY.p2}
+//     color: ${ PALETTE.black };
+//     font-weight: normal;
+//     height: 1.25rem;
+//     left: 1rem;
+//     opacity: ${props => props.disabled ? "24%": "48%"};
+//     pointer-events: none;
+//     position: absolute;
+//     text-align: center;
+//     top: calc(50% - (1.25rem / 2));
+// `;
+
 const StyledInputContainer = styled.div<{disabled : boolean, isError : boolean, isSuccess : boolean, readOnly : boolean}>`
     border-radius: ${LAYOUT.borderRadius};
     height: 3.5rem;
@@ -55,20 +68,6 @@ const StyledDescription = styled.p`
 const StyledErrorMessage = styled(StyledDescription)`
     color: ${PALETTE.red};
 `;
-
-export const StyledLabel = styled.label<{disabled : boolean}>`
-    ${TYPOGRAPHY.p2}
-    color: ${ PALETTE.black };
-    font-weight: normal;
-    height: 1.25rem;
-    left: 1rem;
-    opacity: ${props => props.disabled ? "24%": "48%"};
-    pointer-events: none;
-    position: absolute;
-    text-align: center;
-    top: calc(50% - (1.25rem / 2));
-`;
-
 
 interface I_InputContainerProps{
     description? : string,
