+++ components/InputText.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { InputContainer, StyledLabel, Icon, IconMediumId } from './';
+import { InputContainer, Icon, IconMediumId } from './';
 import { PALETTE, TYPOGRAPHY } from '../utils/Theme';
+import { StyledLabel } from './StyledLabel';
 
 const StyledTextInput = styled.input.attrs({ type: "text" })`
     ${TYPOGRAPHY.p2}
