+++ components/SelectTags.tsx
@@ -3,7 +3,6 @@ import styled, {css} from 'styled-components';
 
 import {    SelectWrapper, I_SelectWrapperProps, SelectOptionDataType, multiSelectionFunctions, multiSelectionProps,
             Tag, TagColor, TagSize,
-            StyledLabel,
             } from './';
 import { PALETTE, TYPOGRAPHY } from '../utils/Theme';
 import { StyledScreenReaderOnly } from '../utils/utils';
@@ -11,6 +10,8 @@ import useTagOverflowCheck, { StyledTagsContainer as StyledTagsContainer_Base} f
 import useUpdatingRef from '../utils/UseUpdatingRef';
 import useOptionsList from '../utils/UseOptionsList';
 import useFocusMonitor from '../utils/UseFocusMonitor';
+
+import { StyledLabel } from './StyledLabel';
 
 const StyledCentreLabel = styled(StyledLabel)`
     top: calc(50% - (1.25rem / 2) - 0.125rem);
