+++ components/contextMenu/ContextMenu.tsx
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import styled from 'styled-components';
 
-import {SHADOW, PALETTE, LAYOUT} from '../utils/Theme';
+import {SHADOW, PALETTE, LAYOUT} from '../../utils/Theme';
 
 // This is exported for use in ContextMenu-related Stories, so you don't need to right-click every time
 export const StyledContextMenu = styled.div<{x : number, y: number}>`
