+++ components/index.tsx
@@ -1,3 +1,6 @@
+export * from './contextMenu/';
+export * from './icons/';
+
 export * from './Avatar';
 export {default as Avatar} from './Avatar';
 export * from './Badge';
@@ -8,14 +11,6 @@ export * from './Checkbox';
 export {default as Checkbox} from './Checkbox';
 export * from './CircleAroundIcon';
 export {default as CircleAroundIcon} from './CircleAroundIcon';
-export * from './ContextMenu';
-export {default as ContextMenu} from './ContextMenu';
-export * from './ContextMenuCheckbox';
-export {default as ContextMenuCheckbox} from './ContextMenuCheckbox';
-export * from './ContextMenuFieldset';
-export {default as ContextMenuFieldset} from './ContextMenuFieldset';
-export * from './ContextMenuRadio';
-export {default as ContextMenuRadio} from './ContextMenuRadio';
 export * from './Files';
 export {default as Files} from './Files';
 export * from './Filter';
@@ -24,18 +19,6 @@ export * from './Grid';
 export {default as Grid} from './Grid';
 export * from './Heading';
 export {default as Heading} from './Heading';
-export * from './Icon';
-export {default as Icon} from './Icon';
-export * from './IconMediumSubject';
-export {default as IconMSubject} from './IconMediumSubject';
-export * from './IconsExtraLarge';
-export {default as IconXL} from './IconsExtraLarge';
-export * from './IconsLarge';
-export {default as IconL} from './IconsLarge';
-export * from './IconsMedium';
-export {default as IconM} from './IconsMedium';
-export * from './IconsSmall';
-export {default as IconS} from './IconsSmall';
 export * from './InputContainer';
 export {default as InputContainer} from './InputContainer';
 export * from './InputTags';
@@ -65,6 +48,8 @@ export {default as SelectTags} from './SelectTags';
 export * from './SelectText';
 export {default as SelectText} from './SelectText';
 export * from './SelectWrapper';
+export { StyledLabel } from './StyledLabel';
+// export {default as StyledLabel} from './StyledLabel';
 export * from './Styled_InputAndOptions';
 export * from './Tab';
 export {default as Tab} from './Tab';
